Remove App JS Folder.
In favour of App Frontend folder. This makes a bit more sense and is a
great place to relegate a lot of that upfront code to consume that SPA
life. Doubtful in this repo but something that would likely be a reality
if this had been a pure API like experience. Maybe ELM?